tenders-llm/requirements.txt
pandas==2.2.2
pyarrow==17.0.0
numpy==2.0.2
scikit-learn==1.5.2
langdetect==1.0.9
tqdm==4.66.5
tenacity==9.0.0
python-dotenv==1.0.1
openai==1.51.2

